Add SearchBar + Autocomplete components
Autocomplete was implemented by passing a ref
to an input element. This way, we can extend
this Autocomplete component to other "SearchBar"
like components, similar to how it is done in the SDK.

Currently the SearchBar only handles vertical searches,
since there is no UniversalResults component yet.
It should be very easy to just call executeUniversalQuery()
if verticalKey is null, though.

These changes depend on my local changes to answers-headless
that return the answers-core requests through stateful core,
instead of hiding them.

TEST=manual

test clicking in and out of the autocomplete, and selecting
with arrow keys