Add the `VerticalResultsHeader` component.
This component will render information about the result set, such as applied query filters, result count, etc.
By using the `props.children` pattern, we do not need to pass configuration for this component to the
`VerticalResults` parent component.

Later PRs will expand on the actual information included in the header.