Alternative Verticals Component (#17)
Add a component for no results / alternative verticals

This component will display a message for no results and indicate that all results are displaying instead. The component will also provide links to view the search on alternative verticals or to view the search across all verticals. I updated the vertical results component to show all results if no results are returned.
 
The links to click the alternative verticals work by changing the vertical key and executing a new search. The 'view results across all verticals' properly fires off a universal search, however we don't have a universal results component yet so the page doesn't change. Also, the vertical suggestion icon customization seemed out of scope so I didn't include it in this PR.

J=1511
TEST=manual

Build the site and test a query which returns no results and alternative verticals. See the suggestions appear and click the vertical link. See that a new search is fired off and see the results appear. Confirm that a universal query is fired when clicking 'view results across all verticals'. See all results appear on the page when no results are returned from the current vertical.