Create generic InputDropdown component (#18)
Create a standalone InputDropdown component and refactor the search bar to use it

The InputDropdown is a controlled component, meaning the state of the input value and the options is maintained outside of the component and passed in as props. This allows answers headless to maintain the state. The InputDropdown controls the input element and the Dropdown component

The behavior is the same as the previous search bar except with a few changes to match the functionality of the SDK/Theme.  When the user uses the keyboard and moves the focus off of one of the options and onto the input element, the input value will return to what had been previously typed in the input box. Additionally, when a user clicks away from the component and then back into the input container, the dropdown options will appear again

J=1539
TEST=manual

Test typing in the search bar, see the autocomplete results appear, test using the arrow keys to navigate the options, test that a search can be performed