Prevent stale listener trigger (#36)
This pr is to address of state update on unmounted component due to stale listeners still in effect on dispatch call. This is a caveat stated in [redux source code](https://github.com/reduxjs/redux/blob/master/src/createStore.ts#L154) in which redux doesn't guarantee unsubscriptions to happen until next dispatch. To solve this, a guard `unsubscribed` is added, which check when component is unmounted and prevent listener's callback from triggering a new render.

TEST=manual

smoke test with universalresults pr where query different searches that previously would generate error
added an integration test for useAnswersState that check for this error when executing multiple searches when nested component structure, see that test pass with the `unsubscribed` guard. Remove the guard and see that test failed.