Add autocomplete screen reader support (#48)
Add a `ScreenReader` component that is used to provide screen reader instructions and announcements. Use this component in `InputDropdown` to give autocomplete instructions and screen reader announcements for number of autocomplete options found.

J=SLAP-1402
TEST=manual

Check that screen reader instructions and announcements are correct in the sample app. An announcement is made for each input change and on clicking the input bar if there is input present.