Add labels to applied static filters (#51)
Allow users to specify group labels for applied static filters. `DecoratedAppliedFilters` accepts a mapping of `fieldId` to group label. If no label is provided, it defaults back to using the `fieldId` as the group label.

J=SLAP-1658
TEST=manual

Check that specifying a group label for a static filter fieldId is reflected when a filter with that fieldId is applied