base Visual Autocomplete
This PR tries to implement Visual Autocomplete in a more flexible
way than a huge array of config options. Externally, VisualAutocompleteSections
can be passed as children into a VisualSearchBar component. This allows the sections
to be organized in many more degrees of freedom than config options could allow, as well
as allowing users to insert whatever other random markup they desire. The current design is
based on the recent InputDropdown refactor and also React-Bootstrap's dropdown component.
If we so desire, we can easily spin up different variations of VisualAutocomplete, with more
hardcoded config options and layouts.

Params like universalLimit and restrictedVerticals are calculated from the children passed into
the component, to try to reduce the amount of config options.

Also fixed a (likely unimportant) edge case with near me searches, and refactored some things
I needed to use in VisualSearchBar.

J=SLAP-1709
TEST=manual

checked I can tab through the autocomplete options and visual entities

debouncing works as expected

loading state is applied (did not walk through it with product but should be easy to tinker with later on)

test universalLimit is passed correctly

test that searches will perform an autocomplete request to check for near me search intent when an autocomplete option
is selected. previously, only searches that were directly typed in could trigger near me behavior