Add useComposedCssClasses hook (#61)
Add the ability to compose CSS classes

This hook allows custom CSS classes to be combined in various ways with the default styling of a given component. Each component will have a `composition` prop which will determine how the composition behaves. It is a wrapper around css-modules-theme. This PR doesn't integrate the functionality into any components yet since the use of Tailwind is still up in the air, and that would affect the integration. However once that's decided, this utility will be added to all components.

The function seemed to impact performance a bit, but adding `useMemo` fixes that

J=1639
TEST=manual

Test this functionality in the Searchbar with Tailwind and test that the component's built in styling works as well as custom styling.