Filters Styling (#79)
Add tailwind styling to the StaticFilters and Facets components
 
Add a divider component which is used by both Facets and StaticFilters. I think long term we should refactor the Facets and static Filters so that they are using the same logic for rendering. This would reduce code and styling duplication. I did a small refactor of the `StaticFilters` component so that it now takes in an array of facets. This allows the StaticFilters component to render the divider between each set of Filters.

Add @tailwindcss/forms to style the filter options.

Also comment out the non-styled components in the App.

J=SLAP-1639
TEST=visual

Visually compare the sample app to the mocks