Use answers-js-docs answers experience (#84)
Update the sample app to use the answers-js-docs answers experience

Update to use the new experience by creating one page per vertical. I also refactored some common code into a `usePage` hook to reduce code duplication across pages. 

There appears to be an issue were the verticalKey isn't set properly when navigating between pages, however this issue existed before this PR which suggests that it's not related to the `usePage` refactor.

I also updated the key in the AppliedFilters so that the same filter value of different filter types won't cause a React error. 

When FilterSearch and Facets are both on the page, the applied filters component will render both of the filters. I think we should update the app so that just one of them is rendered, but this is out of scope.

J=SLAP-1742
TEST=manual

Load the page and see results from the new experience