Pass Client-SDK agent to Headless (#110)
Passes the `ANSWERS_HEADLESS_REACT` agent as part of the Client-SDK header to Headless through `provideAnswersHeadless`.

J=SLAP-1951
TEST=manual

Hook up locally to the starter app and see that requests send the Headless React, Headless, and Core agents as part of the Client-SDK header. Requests to the universal search endpoint made for entity previews only send the Headless and Core agents because the hook uses `provideAnswersHeadless` directly to get an instance of Headless instead of going through the `AnswersHeadlessProvider` component from Headless React.