react 16.8 support branch (#127)
Fixes the react and react-dom versions to 16.8
Changes the tsconfig react transform to React.createElement calls, and adds React imports to all
files since that is a requirement for the old jsx transform.

Update unit tests to pass, they were complaining that we shouldn't be awaiting an "act" call and
the tests were failing. Unfortunately async "act" calls were introduced in react 16.9 so some tests had
to be changed. One test had to be deleted, since it is now impossible to test in the same way and would
need to be completely rewritten. Since this is a one-off build, and react 16.8 users probably aren't knashing
their teeth over extra renders (which we shouldn't have any issues with unless we make a variety of large changes)
I just removed the test.

In useReducer, an argument is always required in react 16.8, even if you manually specify
the generic, so I passed it null.

J=SLAP-2129
TEST=manual

use create-react-app to create a react 16.8 test site
install my local answers-headless-react using `npm install ~/answers-headless-react --install-links`

Add a react 16.8 test site with simple answers components