Rework the `SearchHeadlessProvider` to be a pass-through. (#156)
Per Product's request, the `SearchHeadlessProvider` is now a mere pass-through to the `SearchHeadlessContext.Provider`. Users will need to create and initialize a `Headless` instance themselves. This PR also removed the `Client-SDK` header for Headless React. Product and I determined that it was not useful to track the Headless React version. 

The `acquireSessionId` utility will be moved to either the Component Library or Headless, TBD. The Jest test that was added ensures the `Provider` does not touch anything in the `Headless` instance.

TEST=auto

Ran all unit tests.