PagesJS compatibility (#190)
Introduce tsup and improve the package.json file in order to support PagesJS.

This works in both dev and prod mode of pagesJS

J=BACK-2545
TEST=manual

Publish an alpha version and test it with https://github.com/yext/search-ui-react/pull/400 and @yext/pages@1.0.0-rc.3. Add several search components to the page and confirm they work as expected.